web/pyproject.toml
[tool.poetry]
name = "citizens"
version = "0.1.0"
description = ""
authors = ["Aleksey <Ignatenko.A.E@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
flask_migrate = "^2.5"
flask_script = "^2.0"
flask_sqlalchemy = "^2.4"
psycopg2 = "^2.8"
uwsgi = "^2.0"
ujson = "^1.35"
PyHamcrest = "^1.9"
numpy = "^1.17"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
